+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 -   Added emoji variation selectors to `PhraseView` to ensure proper rendering of color emoji on WebKit browsers.
 -   Empty list values should have an undefined item type, not a never type.
 -   Fixed types of Webpage stream, should have been number, not none.
+-   Don't show project footer in tutorial when in non-editable mode.
 -   [#410](https://github.com/wordplaydev/wordplay/issues/410): Fixed alignment of project preview glyphs.
 
 ### Maintenance